Add tests file which runs the examples as part of the tests.
This is useful because our test suite is not very complete and will not
be during prototyping since we don't have time.

This commit also makes sure the tests run in a reasonable amount of
time: 5s per test on my desktop. I also made inner (chosen
arbitrarily) use cpucores to test it.