Allow to colocate tasks on the same device
The previous version places one task at a time on each device.
If any task is run on the device, it grabs a logical lock
to block placing other tasks on that.
This design was a baseline for a thread pool and was not efficient
since it did not fully exploit computation/memory power.
This commit replaces this with resource aware manner.
Scheduler checks if the target device has enough resources and if it is,
allow to colocate multiple tasks on the same device.